+++ inference/main_cifar-10.cpp
@@ -150,8 +150,8 @@ int main(int argc, char *argv[])
         clock_t run_time_start = clock();
 
 		MatrixXd conv1;
-		double gemm_time_1;
-		double offline_time_1;
+		float gemm_time_1;
+		float offline_time_1;
 		std::tie(conv1, gemm_time_1, offline_time_1) = convolve(image, im_size_1, im_height_1, im_width_1, im_depth_1, k_size_1, stride_1, conv1_b, p1_1, p2_1, conv1_w, output_size_1, mode);
 		
 		MatrixXd pool1 = pool(conv1, f_1, s_1, output_width_1, output_height_1, pp1_1, pp2_1, mode_1);
@@ -159,8 +159,8 @@ int main(int argc, char *argv[])
 		MatrixXd relu1 = relu(pool1);
 		
 		MatrixXd conv2;
-		double gemm_time_2;
-		double offline_time_2;
+		float gemm_time_2;
+		float offline_time_2;
 		std::tie(conv2, gemm_time_2, offline_time_2) = convolve(relu1, im_size_2, im_height_2, im_width_2, im_depth_2, k_size_2, stride_2, conv2_b, p1_2, p2_2, conv2_w, output_size_2, mode);
 		
 		MatrixXd relu2 = relu(conv2);
@@ -168,8 +168,8 @@ int main(int argc, char *argv[])
 		MatrixXd pool2 = pool(relu2, f_2, s_2, output_width_2, output_height_2, pp1_2, pp2_2, mode_2);
 		
 		MatrixXd conv3;
-		double gemm_time_3;
-		double offline_time_3;
+		float gemm_time_3;
+		float offline_time_3;
 		std::tie(conv3, gemm_time_3, offline_time_3) = convolve(pool2, im_size_3, im_height_3, im_width_3, im_depth_3, k_size_3, stride_3, conv3_b, p1_3, p2_3, conv3_w, output_size_3, mode);
 		
 		MatrixXd relu3 = relu(conv3);
@@ -182,7 +182,7 @@ int main(int argc, char *argv[])
 		
         clock_t run_time_end = clock();
 
-        double run_time = (double) (run_time_end-run_time_start) / CLOCKS_PER_SEC;
+        float run_time = (float) (run_time_end-run_time_start) / CLOCKS_PER_SEC;
 		run_time_total += (run_time - offline_time_1 - offline_time_2 - offline_time_3);
 		gemm_time_total += 0.0 + gemm_time_1 + gemm_time_2 + gemm_time_3;
