+++ layers/convolution_layer/convolution.cpp
@@ -69,7 +69,7 @@ MatrixXd im2col(const MatrixXd &input, const int k_size, const int stride) {
     return result;
 }
 
-std::tuple<MatrixXd, double, double> convolve(const MatrixXd &image, const int im_size, const int im_height, const int im_width, const int im_depth, const int k_size, const int stride, const VectorXd &b, const int p1, const int p2, const MatrixXd &w, const int output_size, std::string mode) {
+std::tuple<MatrixXd, float, float> convolve(const MatrixXd &image, const int im_size, const int im_height, const int im_width, const int im_depth, const int k_size, const int stride, const VectorXd &b, const int p1, const int p2, const MatrixXd &w, const int output_size, std::string mode) {
     // im2col for each slice, then concatinate slices.
     MatrixXd im(output_size, k_size*im_depth);
     for (int d=0; d < im_depth ; d++) {
@@ -90,8 +90,8 @@ std::tuple<MatrixXd, double, double> convolve(const MatrixXd &image, const int i
         clock_t start = clock();
         MatrixXd c = im*w.transpose();
         clock_t end = clock();
-        double gemm_time = (double) (end-start) / CLOCKS_PER_SEC;
-        double offline_time = 0;
+        float gemm_time = (float) (end-start) / CLOCKS_PER_SEC;
+        float offline_time = 0;
         
         // Add biases.
         c.rowwise() += b.transpose();
@@ -105,8 +105,8 @@ std::tuple<MatrixXd, double, double> convolve(const MatrixXd &image, const int i
     if (mode == "gemmlowp") {
         // GEMM Multiplication Operation w/ Gemmlowp.
         MatrixXd c;
-        double gemm_time;
-        double offline_time;
+        float gemm_time;
+        float offline_time;
         std::tie(c, gemm_time, offline_time) = glp(im.rows(), im.cols(), w.transpose().cols(), im, w.transpose());
         
         // Add biases.
